sdks/python/stepflow-api/pyproject.toml
[project]
name = "stepflow-api"
version = "0.1.0"
description = "Generated API client for Stepflow server"
requires-python = ">=3.11"
dependencies = [
    "httpx>=0.25",
    "attrs>=22.2.0",
    "python-dateutil>=2.8.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/stepflow_api"]

[tool.ruff]
line-length = 120

[tool.ruff.lint]
select = ["F", "I", "UP"]
